component storage issue diagnosed and fixed
it was because in the handling of removing components, we would swap
components within the storage. then we would update the entity's
componentid list so that the removed component ref would be removed, and
that the component ref pointing to the item that was swapped be updated
to point at the new correct index within storage. this part was fine,
but when updating the component id refs, we forgot to check that the
componentid component storage type (base, updatable, renderable) when
finding the componentid pointing to the moved component in storage.